src/learning/spiking_jelly_module.py
import torch
import torch.nn as nn
import spikingjelly as sj
from spikingjelly import visualizing
from matplotlib import pyplot as plt
import os 
import time
from spikingjelly.activation_based import neuron, functional, layer, monitor, surrogate
from spikingjelly.activation_based import layer, learning
import zlib
import torch.nn.functional as F
from torch.optim import SGD, Adam
from spikingjelly.activation_based.cuda_kernel import tensor_cache


#interface
input_size = 0
output_size = 0
hidden_layer_width = 0
hidden_layer_height = 0
time_interval = 0
threshold = 0
nap_time = 0
epochs = 0
batch_size = 0
neuron_capacity = 0

##apparently there is  a stance regaridng SNN which should be updated with states 

##



# network initialization
state = torch.zeros(1, output_size) # state
w = torch.zeros(input_size, output_size) # weights

# Data Stream 
class Data_stream():
    def __init__(self):
        self.input_data = torch.randn(1, input_size)
        #scaling the data input to be in the range of 0-1 for better training stability

    def __iter__(self):
        yield self.input_data # In practice, this would yield batches of data over time

    def train_data(self):
        # This method can be used to update the data stream with new data during training
        pass

    def test_data(self):
        # This method can be used to update the data stream with new data during testing
        pass

class SNNLayer():
    def __init__(self, input_size, output_size, device, threshold=1.0):
        self.input_size = input_size
        self.output_size = output_size
        self.device = device
        self.threshold = threshold

        # Initialize weights and state variables on the GPU
        self.weights = torch.randn(input_size, output_size, device=device) * 0.1
        self.voltages = torch.zeros(output_size, device=device)
        self.spikes = torch.zeros(output_size, device=device)
    
    def forward(self, input):
        # Call the CUDA kernel for the forward pass
        snn_cuda.snn_forward(input, self.voltages, self.spikes, self.weights)
        return self.spikes
    
    def reset(self):
        if neuron_capacity > threshold:
            self.voltages.zero_()
            self.spikes.zero_()
            self.time.sleep(nap_time) # simulate nap time for energy efficiency 

    def loss(self, output, target):
        # Example loss function (e.g., mean squared error)
        return F.mse_loss(output, target)
    
    def backward(self, grad_output):
        # Implement backward pass (e.g., using surrogate gradients)
        pass

    def update_weights(self, grad_output):
        # Update weights based on gradients and learning rate
        pass

    def gradient_clipping(self, max_norm):
        # Clip gradients to prevent exploding gradients
        pass

    def heartbeat(self):
        # This function can be called periodically to monitor the health of each neuron during operations 

        pass


training = Data_stream.train_data()
testing = Data_stream.test_data()

def train_snn(model, data, optimizer, loss_fn, device):
    model.train()
    for epoch in range(epochs):
        for input in data:
            input = input.to(device)
            output = model.forward(input)
            target = torch.zeros_like(output) # Placeholder target
            loss = loss_fn(output, target)
            optimizer.zero_grad()
            loss.backward()
            model.gradient_clipping(max_norm=1.0) # Example max norm for clipping
            model.update_weights(optimizer)
        print(f"Epoch {epoch+1}/{epochs}, Loss: {loss.item():.4f}")


def test_snn(model, data, device):
    model.eval()
    with torch.no_grad():
        for input in data:
            input = input.to(device)
            output = model.forward(input)
            print(f"Output: {output.cpu().numpy()}")
